+++ bin/processes.py
@@ -197,7 +197,7 @@ def concat_phanotate_meta(out_dir, num_fastas):
                 outfile.write(infile.read())
 
 
-def run_trnascan_meta(filepath_in, out_dir, threads, num_fastas):
+def run_trnascan_meta(filepath_in, out_dir, threads, num_fastas, trna_scan_model):
     """
     Runs trnascan to output gffs one contig per thread
     :param filepath_in: input filepath
@@ -210,12 +210,17 @@ def run_trnascan_meta(filepath_in, out_dir, threads, num_fastas):
     input_tmp_dir = os.path.join(out_dir, "input_split_tmp")
     commands = []
 
+    if trna_scan_model == "general":
+        model = "G"
+    else: # bacterial
+        model = "B"
+
     for i in range(1, num_fastas + 1):
         in_file = "input_subprocess" + str(i) + ".fasta"
         out_file = "trnascan_tmp" + str(i) + ".gff"
         filepath_in = os.path.join(input_tmp_dir, in_file)
         filepath_out = os.path.join(input_tmp_dir, out_file)
-        cmd = "tRNAscan-SE " + filepath_in + " --thread 1 -G -Q -j " + filepath_out
+        cmd = f"tRNAscan-SE {filepath_in} --thread 1 -{model} -Q -j {filepath_out}"
         commands.append(cmd)
 
     n = int(threads)  # the number of parallel processes you want
@@ -632,7 +637,7 @@ def translate_fastas(out_dir, gene_predictor, coding_table, genbank_file):
     # for genbank do nothing
 
 
-def run_trna_scan(filepath_in, threads, out_dir, logdir):
+def run_trna_scan(filepath_in, threads, out_dir, logdir, trna_scan_model):
     """
     Runs trna scan
     :param filepath_in: input filepath
@@ -643,11 +648,16 @@ def run_trna_scan(filepath_in, threads, out_dir, logdir):
 
     out_gff = os.path.join(out_dir, "trnascan_out.gff")
 
+    if trna_scan_model == "general":
+        model = "G"
+    else: # bacterial
+        model = "B"
+
     trna = ExternalTool(
         tool="tRNAscan-SE",
         input=f"{filepath_in}",
         output=f"{out_gff}",
-        params=f"--thread {threads} -G -Q -j",
+        params=f"--thread {threads} -{model} -Q -j",
         logdir=logdir,
         outfile="",
     )